Fix public api issues
